tests: move tx_cache impl
