+++ tests/device_tests/test_msg_signtx.py
@@ -19,9 +19,9 @@ import pytest
 from trezorlib import btc, messages as proto
 from trezorlib.tools import H_, CallException, btc_hash, parse_path
 
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
 from .conftest import TREZOR_VERSION
+from .tx_cache import tx_cache
 
 TXHASH_157041 = bytes.fromhex(
     "1570416eb4302cf52979afd5e6909e37d8fdd874301f7cc87e547e509cb1caa6"
