+++ tests/device_tests/test_msg_signtx_capricoin.py
@@ -19,8 +19,8 @@ import pytest
 from trezorlib import btc, messages
 from trezorlib.tools import parse_path
 
-from ..support.tx_cache import tx_cache
 from .conftest import setup_client
+from .tx_cache import tx_cache
 
 TXHASH_3bf506 = bytes.fromhex(
     "3bf506c81ce84eda891679ddc797d162c17c60b15d6c0ac23be5e31369e7235f"
