+++ tests/device_tests/test_msg_signtx_dash.py
@@ -17,8 +17,8 @@
 from trezorlib import btc, messages as proto
 from trezorlib.tools import parse_path
 
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
+from .tx_cache import tx_cache
 
 TX_API = tx_cache("Dash")
