+++ tests/device_tests/test_msg_signtx_decred.py
@@ -19,8 +19,8 @@ import pytest
 from trezorlib import btc, messages as proto
 from trezorlib.tools import parse_path
 
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
+from .tx_cache import tx_cache
 
 TX_API = tx_cache("Decred Testnet")
