+++ tests/device_tests/test_multisig.py
@@ -21,8 +21,8 @@ from trezorlib.tools import CallException, parse_path
 
 from ..support import ckd_public as bip32
 from ..support.ckd_public import deserialize
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
+from .tx_cache import tx_cache
 
 TX_API = tx_cache("Bitcoin")
