+++ tests/device_tests/test_multisig_change.py
@@ -18,8 +18,8 @@ from trezorlib import btc, messages as proto
 from trezorlib.tools import H_, parse_path
 
 from ..support import ckd_public as bip32
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
+from .tx_cache import tx_cache
 
 TX_API = tx_cache("Testnet")
