+++ tests/device_tests/test_protection_levels.py
@@ -18,8 +18,8 @@ import pytest
 
 from trezorlib import btc, debuglink, device, messages as proto, misc
 
-from ..support.tx_cache import tx_cache
 from .common import TrezorTest
+from .tx_cache import tx_cache
 
 TXHASH_d5f65e = bytes.fromhex(
     "d5f65ee80147b4bcc70b75e4bbf2d7382021b871bd8867ef8fa525ef50864882"
